Split env parsing into app/env.ts
Agent-Logs-Url: https://github.com/mcansh/remix-wordle/sessions/841a4215-24bd-4908-9c77-51c5cfeab87e
