+++ app/utils/db.ts
@@ -1,7 +1,7 @@
 import { PrismaPg } from "@prisma/adapter-pg"
 import Redis from "ioredis"
 
-import { env } from "#app/constants.ts"
+import { env } from "#app/env.ts"
 import { PrismaClient } from "#app/generated/prisma/client.ts"
 
 const adapter = new PrismaPg({ connectionString: process.env.DATABASE_URL })
